Solid-JS For component fix (#7)
<!-- 
Thanks for creating this pull request 🤗

Please make sure that the pull request is limited to one type (docs,
feature, etc.) and keep it as small as possible. You can open multiple
prs instead of opening a huge one.
-->

<!-- If this pull request closes an issue, please mention the issue
number below -->
Closes #6

## 📑 Description
Wraps the for component in a conditional to not attempt to render it if
the list is empty

<!-- You can also choose to add a list of changes and if they have been
completed or not by using the markdown to-do list syntax
- [ ] Not Completed
- [x] Completed
-->

## ✅ Checks
<!-- Make sure your pr passes the CI checks and do check the following
fields as needed - -->
- [x] My pull request adheres to the code style of this project
- [x] My code requires changes to the documentation
- [x] I have updated the documentation as required
- [x] All the tests have passed

## ℹ Additional Information
<!-- Any additional information like breaking changes, dependencies
added, screenshots, comparisons between new and old behavior, etc. -->

(This did not happen in local dev)